+++ mkart/services/lap.py
@@ -1,3 +1,4 @@
+import operator
 from functools import reduce
 
 
@@ -18,11 +19,17 @@ class LapService:
 
     def _get_race_duration_of_driver(self, laps):
         durations_of_laps = [lap.duration for lap in laps]
-        return reduce((lambda x, y: x + y), durations_of_laps)
+        return reduce(operator.add, durations_of_laps)
 
     def _get_best_lap(self, laps):
         laps.sort(key=lambda lap: lap.duration)
         return laps[0]
+
+    def _get_average_speed_of_laps(self, laps):
+        len_laps = len(laps)
+        average_speed_laps = [lap.average_speed for lap in laps]
+        total = reduce(operator.add, average_speed_laps)
+        return total / len_laps
 
     def get_last_laps_duration(self):
         last_laps = []
@@ -50,3 +57,15 @@ class LapService:
 
     def get_best_lap_of_race(self):
         return self._get_best_lap(self._laps)
+
+    def get_drivers_average_speed(self):
+        drivers_average_speed = []
+        drivers_ids = self._get_drivers_id()
+        for id in drivers_ids:
+            driver_laps = self._get_driver_laps(id)
+            driver = driver_laps[0].driver
+            driver.average_speed = self._get_average_speed_of_laps(driver_laps)
+            drivers_average_speed.append(driver)
+
+        drivers_average_speed.sort(key=lambda x: x.average_speed, reverse=True)
+        return drivers_average_speed